+++ classy/index.py
@@ -1,4 +1,5 @@
 """Module to manage the global spectra index in classy."""
+import sys
 from datetime import datetime
 from functools import cache
 
@@ -24,9 +25,10 @@ def load():
         The global spectra index. Empty if index does not exist yet.
     """
     if not PATH.is_file():
-        logger.error(
-            f"No reflectance spectra are available. Run '$ classy status' to retrieve them."
-        )
+        if "status" not in sys.argv:
+            logger.error(
+                f"No reflectance spectra are available. Run '$ classy status' to retrieve them."
+            )
         return pd.DataFrame(data={"name": [], "source": [], "filename": []}, index=[])
 
     return pd.read_csv(PATH, dtype={"number": "Int64"}, low_memory=False)
@@ -47,7 +49,12 @@ def add(entries):
     entries : list of pd.DataFrame
         The entries to add to the index.
     """
-    index = load()
+
+    # Cannot use the load() function here due to caching
+    if not PATH.is_file():
+        index = pd.DataFrame(data={"name": [], "source": [], "filename": []}, index=[])
+    else:
+        index = pd.read_csv(PATH, dtype={"number": "Int64"}, low_memory=False)
 
     # Find overlap between new entries and exisiting index
     # Store the DF index of the classy index to later drop duplicates via that index
@@ -61,7 +68,6 @@ def add(entries):
         index = index.drop(overlap.index)
 
     index = pd.concat([index, entries], ignore_index=True)
-
     save(index)